定时任务更新推荐

+++ src/main/java/com/zzx/zzx_music_recommendation_system/entity/Recommend.java
@@ -37,7 +37,7 @@ public class Recommend implements Serializable {
     private Long musicId;
 
     @TableField("socre")
-    private Integer socre;
+    private Float socre;
 
     @TableField("recommend_type")
     private Integer recommendType;
